some changes 4

+++ src/main/java/com/sport/service/services/SubscriberService.java
@@ -17,7 +17,7 @@ public interface SubscriberService {
 
     void updateSubscriber(Subscriber subscriber, Long id);
 
-    int getUsersCount();
+    long getUsersCount();
 
     int getSubscriptionsCount();
